+++ src/components/WalletConnect/v1/Connected.tsx
