+++ src/components/WalletConnect/v1/Disconnected.tsx
