+++ src/components/WalletConnect/v1/PendingConnection.tsx
