+++ src/components/WalletConnect/v1/PendingRequest.tsx
