+++ src/components/WalletConnect/v1/QRScanner.tsx
